refactor: move key handling to own file test: cover new functionality with unit tests
